apps/platform-api/src/agents/image_agent/requirements_web.txt
Flask==2.3.3
diffusers>=0.25.0
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.1
accelerate>=0.24.1
huggingface-hub>=0.19.0